+++ packages/cedar/CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
-- bump @esri/arcgis-rest-js dependencies to 1.7.0
+- bump @esri/arcgis-rest-js dependencies to 1.7.1
 
 ## [1.0.0-beta.8]
 ### Added